Move extensions and types from Mappers and Model into feature folders
