+++ apps/dh/api-dh/source/DataHub.WebApi/Modules/Common/Enums/MeteringPointType.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Energinet.DataHub.WebApi.Model.MeteringPoint;
+namespace Energinet.DataHub.WebApi.Modules.Common.Enums;
 
 public enum MeteringPointType
 {
