+++ apps/dh/api-dh/source/DataHub.WebApi/Modules/Common/Enums/Resolution.cs
@@ -12,10 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Energinet.DataHub.WebApi.Model.AggregatedMeasureData;
+namespace Energinet.DataHub.WebApi.Modules.Common.Enums;
 
-public enum SettlementMethod
+public enum Resolution
 {
-    NonProfiled,
-    Flex,
+    Monthly,
+    Hourly,
+    QuarterHourly,
 }
