+++ apps/dh/api-dh/source/DataHub.WebApi/Modules/ElectricityMarket/Measurements/MeasurementOperations.cs
@@ -22,10 +22,10 @@ using Energinet.DataHub.ProcessManager.Abstractions.Api.SendMeasurements.Model;
 using Energinet.DataHub.ProcessManager.Client;
 using Energinet.DataHub.WebApi.Common.Exceptions;
 using Energinet.DataHub.WebApi.Extensions;
-using Energinet.DataHub.WebApi.Mapper.Measurement;
-using Energinet.DataHub.WebApi.Model.Measurements;
 using Energinet.DataHub.WebApi.Modules.ElectricityMarket.Extensions;
+using Energinet.DataHub.WebApi.Modules.ElectricityMarket.Mappers;
 using Energinet.DataHub.WebApi.Modules.ElectricityMarket.Measurements.Extensions;
+using Energinet.DataHub.WebApi.Modules.ElectricityMarket.Measurements.Types;
 using Energinet.DataHub.WebApi.Modules.RevisionLog.Attributes;
 using HotChocolate.Authorization;
 using NodaTime;
