+++ apps/dh/api-dh/source/DataHub.WebApi/Modules/ElectricityMarket/Measurements/Types/Measurement.cs
@@ -12,11 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Energinet.DataHub.WebApi.Model;
+namespace Energinet.DataHub.WebApi.Modules.ElectricityMarket.Measurements.Types;
 
-public enum SettlementVersion
-{
-    FirstCorrection,
-    SecondCorrection,
-    ThirdCorrection,
-}
+public record Measurement(
+    int Position,
+    double Quantity,
+    Quality Quality);
