+++ apps/dh/api-dh/source/DataHub.WebApi/Modules/ElectricityMarket/Measurements/Types/SendMeasurementsQualityType.cs
@@ -12,8 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.WebApi.Model.Measurements;
-
 namespace Energinet.DataHub.WebApi.Modules.ElectricityMarket.Measurements.Types;
 
 public class SendMeasurementsQualityType : EnumType<Quality>
