+++ apps/dh/api-dh/source/DataHub.WebApi/Modules/Processes/Requests/Extensions/SettlementMethodExtensions.cs
@@ -12,25 +12,24 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.EDI.B2CClient.Abstractions.RequestWholesaleSettlement.V1;
-using Energinet.DataHub.WebApi.Model;
+using Energinet.DataHub.EDI.B2CClient.Abstractions.RequestAggregatedMeasureData.V1;
+using Energinet.DataHub.WebApi.Modules.Processes.Requests.Models;
 
-namespace Energinet.DataHub.WebApi.Mapper;
+namespace Energinet.DataHub.WebApi.Modules.Processes.Requests.Extensions;
 
-public static class ChargeTypeExtensions
+public static class SettlementMethodExtensions
 {
-    public static ChargeTypeV1 MapToRequestWholesaleSettlementV1(this ChargeType source)
+    public static SettlementMethodV1 MapToRequestAggregatedMeasureDataV1(this SettlementMethod source)
     {
         return source switch
         {
-            ChargeType.Fee => ChargeTypeV1.Fee,
-            ChargeType.Subscription => ChargeTypeV1.Subscription,
-            ChargeType.Tariff => ChargeTypeV1.Tariff,
+            SettlementMethod.Flex => SettlementMethodV1.Flex,
+            SettlementMethod.NonProfiled => SettlementMethodV1.NonProfiled,
         };
     }
 
-    public static ChargeTypeV1? MapToRequestWholesaleSettlementV1(this ChargeType? source)
+    public static SettlementMethodV1? MapToRequestAggregatedMeasureDataV1(this SettlementMethod? source)
     {
-        return source?.MapToRequestWholesaleSettlementV1();
+        return source?.MapToRequestAggregatedMeasureDataV1();
     }
 }
