+++ apps/dh/api-dh/source/DataHub.WebApi/Modules/Processes/Requests/Models/ChargeType.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Energinet.DataHub.WebApi.Model;
+namespace Energinet.DataHub.WebApi.Modules.Processes.Requests.Models;
 
 public enum ChargeType
 {
