+++ apps/dh/api-dh/source/DataHub.WebApi/Modules/Processes/Requests/Models/SettlementMethod.cs
@@ -12,11 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Energinet.DataHub.WebApi.Model;
+namespace Energinet.DataHub.WebApi.Modules.Processes.Requests.Models;
 
-public enum Resolution
+public enum SettlementMethod
 {
-    Monthly,
-    Hourly,
-    QuarterHourly,
+    NonProfiled,
+    Flex,
 }
