+++ apps/dh/api-dh/source/DataHub.WebApi/Modules/Processes/Requests/Models/SettlementVersion.cs
@@ -12,10 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Energinet.DataHub.WebApi.Model.Measurements;
+namespace Energinet.DataHub.WebApi.Modules.Processes.Requests.Models;
 
-public enum MeasurementUnit
+public enum SettlementVersion
 {
-    KilowattHour,
-    KiloVoltAmpereReactiveHour,
+    FirstCorrection,
+    SecondCorrection,
+    ThirdCorrection,
 }
